server/auth-server/requirements.txt
fastapi
uvicorn[standard]
bcrypt
pyjwt
pymongo